Update device discovery logic to be independent of device naming convention
