+++ src/index.ts
@@ -24,13 +24,13 @@ export var OUTPUT_D = "outD";
 
 // IO
 export var Device = io.Device;
+export var IndexedDevice = io.IndexedDevice;
 
 // Motors
 export var Motor = motors.Motor;
 export var DcMotor = motors.DcMotor;
 export var LargeMotor = motors.LargeMotor;
 export var MediumMotor = motors.MediumMotor;
-export var MotorSpeedSp = motors.MotorSpeedSp;
 export var ServoMotor = motors.ServoMotor;
 
 //~autogen export-string-literal-types classes.motor>currentClass "motors">module
