+++ src/io.ts
@@ -286,3 +286,34 @@ export class Device {
         });
     }
 }
+
+export class IndexedDevice extends Device {
+    protected deviceIndexRegex = new RegExp("(\\d+)", "g");
+    
+    protected _deviceIndex: number = -1;
+    get deviceIndex(): number {
+        return this._deviceIndex;
+    }
+
+    constructor(driverTypeDirName: string, nameConvention?: string, targetAddress?: string, targetDriverName?: string | string[]) {
+        super();
+
+        var propertyConstraints: {[propertyName: string]: any} = {};
+
+        if (targetAddress != undefined)
+            propertyConstraints['address'] = targetAddress;
+
+        if (targetDriverName != undefined)
+            propertyConstraints['driver_name'] = [].concat(targetDriverName);
+
+        this.connect(driverTypeDirName, nameConvention, propertyConstraints);
+
+        if (this.connected) {
+            var matches = this.deviceIndexRegex.exec(this.deviceDirName);
+            
+            if (matches != null && matches[0] != undefined) {
+                this._deviceIndex = Number(matches[1]);
+            }
+        }
+    }
+}
