+++ src/sensors.ts
@@ -1,32 +1,10 @@
 ﻿import IO = require('./io');
 import Device = IO.Device;
+import IndexedDevice = IO.IndexedDevice;
 
-export class SensorBase extends Device {
-    protected _deviceIndex: number = -1;
-    get deviceIndex(): number {
-        return this._deviceIndex;
-    }
-
-    constructor(driverTypeDirName: string, nameConvention: string, targetAddress?: string, targetDriverName?: string | string[]) {
-        super();
-
-        var propertyConstraints: { [propertyName: string]: any } = {};
-
-        if (targetAddress != undefined)
-            propertyConstraints['address'] = targetAddress;
-
-        if (targetDriverName != undefined)
-            propertyConstraints['driver_name'] = [].concat(targetDriverName);
-
-        this.connect(driverTypeDirName, nameConvention, propertyConstraints);
-
-        if (this.connected) {
-            var matches = new RegExp(nameConvention).exec(this.deviceDirName);
-
-            if (matches != null && matches[0] != undefined) {
-                this._deviceIndex = Number(matches[1]);
-            }
-        }
+export class SensorBase extends IndexedDevice {
+    constructor(driverTypeDirName: string, nameConvention?: string, targetAddress?: string, targetDriverName?: string | string[]) {
+        super(driverTypeDirName, nameConvention, targetAddress, targetDriverName);
     }
 }
